Split up Exp module
