+++ packages/expo-cli/package.json
@@ -51,6 +51,7 @@
     "@types/js-yaml": "^3.12.2",
     "@types/klaw-sync": "^6.0.0",
     "@types/md5-file": "^4.0.2",
+    "@types/minipass": "^2.2.0",
     "@types/npm-package-arg": "^6.1.0",
     "@types/pngjs": "^3.4.1",
     "@types/progress": "^2.0.3",
@@ -99,6 +100,7 @@
     "leven": "^3.1.0",
     "lodash": "^4.17.19",
     "md5-file": "^5.0.0",
+    "minipass": "2.3.5",
     "npm-package-arg": "6.1.0",
     "ora": "3.4.0",
     "pacote": "^11.1.0",
