+++ packages/xdl/package.json
@@ -52,7 +52,7 @@
     "delay-async": "1.2.0",
     "devcert": "^1.1.0",
     "express": "4.16.4",
-    "form-data": "2.3.2",
+    "form-data": "^2.3.2",
     "freeport-async": "2.0.0",
     "fs-extra": "9.0.0",
     "getenv": "0.7.0",
@@ -68,7 +68,6 @@
     "lodash": "^4.17.19",
     "md5hex": "1.0.0",
     "minimatch": "3.0.4",
-    "minipass": "2.3.5",
     "mv": "2.1.1",
     "ncp": "2.0.0",
     "node-forge": "0.7.6",
@@ -76,7 +75,6 @@
     "p-retry": "4.1.0",
     "p-timeout": "3.1.0",
     "package-json": "6.4.0",
-    "pacote": "11.1.0",
     "pascal-case": "2.0.1",
     "pretty-bytes": "^5.3.0",
     "probe-image-size": "~6.0.0",
@@ -112,12 +110,11 @@
     "@expo/ngrok": "2.4.3",
     "@types/analytics-node": "^3.1.1",
     "@types/concat-stream": "^1.6.0",
-    "@types/form-data": "2.2.0",
+    "@types/form-data": "^2.2.0",
     "@types/fs-extra": "^9.0.1",
     "@types/getenv": "^0.7.0",
     "@types/hapi__joi": "^17.1.2",
     "@types/hashids": "^1.0.30",
-    "@types/minipass": "^2.2.0",
     "@types/ncp": "^2.0.1",
     "@types/node-forge": "^0.8.6",
     "@types/raven": "^2.5.3",
