+++ packages/xdl/src/Project.ts
@@ -50,7 +50,6 @@ import * as DevSession from './DevSession';
 import * as EmbeddedAssets from './EmbeddedAssets';
 import { maySkipManifestValidation } from './Env';
 import { ErrorCode } from './ErrorCode';
-import * as Exp from './Exp';
 import logger from './Logger';
 import { Asset, exportAssetsAsync, publishAssetsAsync } from './ProjectAssets';
 import * as ProjectSettings from './ProjectSettings';
@@ -72,6 +71,7 @@ import { assertValidProjectRoot } from './project/errors';
 import { startTunnelsAsync, stopTunnelsAsync } from './project/ngrok';
 import { writeArtifactSafelyAsync } from './tools/ArtifactUtils';
 import FormData from './tools/FormData';
+import { resolveEntryPoint } from './tools/resolveEntryPoint';
 const MINIMUM_BUNDLE_SIZE = 500;
 
 const treekillAsync = promisify<number, string>(treekill);
@@ -942,7 +942,7 @@ async function buildPublishBundlesAsync(
     },
     platforms.map((platform: Platform) => ({
       platform,
-      entryPoint: Exp.determineEntryPoint(projectRoot, platform),
+      entryPoint: resolveEntryPoint(projectRoot, platform),
       dev: bundleOptions.dev,
     }))
   );
@@ -955,7 +955,7 @@ async function buildPublishBundlesAsync(
 
 // Fetch iOS and Android bundles for publishing
 async function fetchPublishBundlesAsync(projectRoot: string, opts?: PackagerOptions) {
-  const entryPoint = Exp.determineEntryPoint(projectRoot);
+  const entryPoint = resolveEntryPoint(projectRoot);
   const publishUrl = await UrlUtils.constructPublishUrlAsync(
     projectRoot,
     entryPoint,
